DOC: Document correct parameter types.

+++ skimage/_shared/interpolation.pyx
@@ -21,7 +21,7 @@ cdef inline double nearest_neighbour_interpolation(double* image, int rows,
         Input image.
     rows, cols : int
         Shape of image.
-    r, c : int
+    r, c : double
         Position at which to interpolate.
     mode : {'C', 'W', 'R', 'N'}
         Wrapping mode. Constant, Wrap, Reflect or Nearest.
@@ -50,7 +50,7 @@ cdef inline double bilinear_interpolation(double* image, int rows, int cols,
         Input image.
     rows, cols : int
         Shape of image.
-    r, c : int
+    r, c : double
         Position at which to interpolate.
     mode : {'C', 'W', 'R', 'N'}
         Wrapping mode. Constant, Wrap, Reflect or Nearest.
@@ -109,7 +109,7 @@ cdef inline double biquadratic_interpolation(double* image, int rows, int cols,
         Input image.
     rows, cols : int
         Shape of image.
-    r, c : int
+    r, c : double
         Position at which to interpolate.
     mode : {'C', 'W', 'R', 'N'}
         Wrapping mode. Constant, Wrap, Reflect or Nearest.
@@ -185,7 +185,7 @@ cdef inline double bicubic_interpolation(double* image, int rows, int cols,
         Input image.
     rows, cols : int
         Shape of image.
-    r, c : int
+    r, c : double
         Position at which to interpolate.
     mode : {'C', 'W', 'R', 'N'}
         Wrapping mode. Constant, Wrap, Reflect or Nearest.
